src/main/java/katas/katas/AdevintaKata/Event.java
package katas.katas.AdevintaKata;

import java.util.Date;
import java.util.List;

public class Event {

    private String title;
    private String location;
    private int startDate;
    private int endDate;
    //private List<Guest> guestList;

    public Event() {
    }

    public Event(String title, String location, int startDate, int endDate) {
        this.title = title;
        this.location = location;
        this.startDate = startDate;
        this.endDate = endDate;
       // this.guestList = guestList;
    }

    public String getTitle() {
        return title;
    }

    public void setTitle(String title) {
        this.title = title;
    }

    public String getLocation() {
        return location;
    }

    public void setLocation(String location) {
        this.location = location;
    }

    public int getStartDate() {
        return startDate;
    }

    public void setStartDate(int startDate) {
        this.startDate = startDate;
    }

    public int getEndDate() {
        return endDate;
    }

    public void setEndDate(int endDate) {
        this.endDate = endDate;
    }

    /*public List<Guest> getGuestList() {
        return guestList;
    }

    public void setGuestList(List<Guest> guestList) {
        this.guestList = guestList;
    }*/
}
